tests/wpt/tests/css/css-values/clamp-length-invalid.html

<!DOCTYPE html>
<link rel="help" href="https://drafts.csswg.org/css-values-4/#comp-func">
<link rel="author" title="Xiaocheng Hu" href="mailto:xiaochengh@chromium.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../support/parsing-testcommon.js"></script>
<script>
function test_invalid_length(value) {
  test_invalid_value('letter-spacing', value);
}

test_invalid_length('clamp()');
test_invalid_length('clamp( )');
test_invalid_length('clamp(,)');
test_invalid_length('clamp(1px, )');
test_invalid_length('clamp(, 1px)');
test_invalid_length('clamp(1px, 1px)');
test_invalid_length('clamp(1px, , 1px)');
test_invalid_length('clamp(, 1px, 1px)');
test_invalid_length('clamp(1px, 1px, )');
test_invalid_length('clamp(1px, 1px, 1px, )');
test_invalid_length('clamp(1px 1px 1px)');
test_invalid_length('clamp(0, 10rem, 100%)');
test_invalid_length('clamp(none, none, none)');
test_invalid_length('clamp(none, none + none, none)');
test_invalid_length('clamp(none, none + 1, none)');
test_invalid_length('clamp(-none, 1px, 1px)');
test_invalid_length('clamp(10px, none * none, 20px)');
test_invalid_length('clamp(10px, 2px * none, 20px)');
test_invalid_length('clamp(10px, 2px + none, 20px)');
test_invalid_length('clamp(10px + none, 10px / 2px + none, 20px / none)');
test_invalid_length('clamp(10px, -none, 20px)');
test_invalid_length('clamp(10px, 15px, abs(none) * 1px)');
test_invalid_length('clamp(10px, 15px, abs(none * 1px))');
</script>
